by ravi

+++ visog-job-portal-ui/WebContent/resources/js/postJobSeeker.js
@@ -78,12 +78,14 @@
 
 						// HTTP DELETE- delete postJobSeeker by Id
 						$scope.deletePostJobSeeker = function(postJobSeeker) {
+							if(confirm("Are you sure to Delete ?") == true){
 							$http(
 									{
 										method : 'DELETE',
 										url : 'http://localhost:8080/visog-job-portal-api/transaction/postJobSeeker/'
 												+ postJobSeeker.id
 									}).then(_success, _error);
+							}
 						};
 
 						// In case of edit, populate form fields and assign form.id
